IT_NidaRizwan_YT/Day07_Handoffs_StructuredOutput/structured_output/pyproject.toml
[project]
name = "structured-output"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
